… src/AdventOfCode/Point3D.cs
… src/AdventOfCode/Puzzles/Y2021/Day22.cs
… src/AdventOfCode/Vector3Extensions.cs