Use the `Vector3` and `Vector4` structs instead of a hand-rolled `Point` class that can be 3D and 4D.

Is it faster though..?