@martincostello 2022-02-18
Of course day 17 didn't have a benchmark because it's too slow...