@martincostello 2022-02-19
```md
| benchmark                  | mean (microbenchmarks-2021.base) | mean (microbenchmarks-2021.pr) | ratio | allocated (microbenchmarks-2021.base) | allocated (microbenchmarks-2021.pr) | ratio |
| -------------------------- | -------------------------------- | ------------------------------ | ----- | ------------------------------------- | ----------------------------------- | ----- |
| Solve2021(input: Y2021-19) |                            1.7 s |                          1.7 s |  1.00 |                               1.42 GB |                             1.42 GB |  1.00 |                          
| Solve2021(input: Y2021-22) |                         219.4 ms |                       236.4 ms |  1.08 |                              140.8 MB |                            140.8 MB |  1.00 |
```