@martincostello 2022-02-19
Day 17 for 2020 is definitely faster.